[semantic-arc] Move main optimize loop of SemanticARCOptVisitor into its own file.
