+++ lib/SILOptimizer/SemanticARC/SemanticARCOpts.cpp
@@ -12,130 +12,13 @@
 
 #define DEBUG_TYPE "sil-semantic-arc-opts"
 
-#include "OwnershipLiveRange.h"
-#include "OwnershipPhiOperand.h"
 #include "SemanticARCOptVisitor.h"
 #include "Transforms.h"
 
-#include "swift/Basic/LLVM.h"
-
-#include "swift/Basic/BlotSetVector.h"
-#include "swift/Basic/FrozenMultiMap.h"
-#include "swift/Basic/MultiMapCache.h"
-#include "swift/Basic/STLExtras.h"
-#include "swift/SIL/BasicBlockUtils.h"
-#include "swift/SIL/DebugUtils.h"
-#include "swift/SIL/LinearLifetimeChecker.h"
-#include "swift/SIL/MemAccessUtils.h"
-#include "swift/SIL/OwnershipUtils.h"
-#include "swift/SIL/Projection.h"
-#include "swift/SIL/SILArgument.h"
-#include "swift/SIL/SILBuilder.h"
-#include "swift/SIL/SILInstruction.h"
-#include "swift/SIL/SILVisitor.h"
-#include "swift/SILOptimizer/Analysis/PostOrderAnalysis.h"
-#include "swift/SILOptimizer/PassManager/Passes.h"
 #include "swift/SILOptimizer/PassManager/Transforms.h"
-#include "swift/SILOptimizer/Utils/InstOptUtils.h"
-#include "swift/SILOptimizer/Utils/ValueLifetime.h"
-#include "llvm/ADT/SmallPtrSet.h"
-#include "llvm/ADT/SmallVector.h"
-#include "llvm/ADT/Statistic.h"
 
 using namespace swift;
 using namespace swift::semanticarc;
-
-//===----------------------------------------------------------------------===//
-//                               Implementation
-//===----------------------------------------------------------------------===//
-
-bool SemanticARCOptVisitor::optimize() {
-  bool madeChange = false;
-
-  // First process the worklist until we reach a fixed point.
-  madeChange |= processWorklist();
-
-  {
-    // If we made a change, set that we assume we are at fixed point and then
-    // re-run the worklist so that we can
-    // properly seeded the ARC peephole map.
-    ctx.assumingAtFixedPoint = true;
-    SWIFT_DEFER { ctx.assumingAtFixedPoint = false; };
-
-    // Add everything in visitedSinceLastMutation to the worklist so we
-    // recompute our fixed point.
-    drainVisitedSinceLastMutationIntoWorklist();
-
-    // Then re-run the worklist. We shouldn't modify anything since we are at a
-    // fixed point and are just using this to seed the
-    // joinedOwnedIntroducerToConsumedOperands after we have finished changing
-    // things. If we did change something, we did something weird, so assert!
-    bool madeAdditionalChanges = processWorklist();
-    (void)madeAdditionalChanges;
-    assert(!madeAdditionalChanges && "Should be at the fixed point");
-  }
-
-  return madeChange;
-}
-
-bool SemanticARCOptVisitor::processWorklist() {
-  // NOTE: The madeChange here is not strictly necessary since we only have
-  // items added to the worklist today if we have already made /some/ sort of
-  // change. That being said, I think there is a low cost to including this here
-  // and makes the algorithm more correct, visually and in the face of potential
-  // refactoring.
-  bool madeChange = false;
-
-  while (!worklist.empty()) {
-    // Pop the last element off the list. If we were returned None, we blotted
-    // this element, so skip it.
-    SILValue next = worklist.pop_back_val().getValueOr(SILValue());
-    if (!next)
-      continue;
-
-    // First check if this is a value that we have visited since the last time
-    // we erased an instruction. If we have visited it, skip it. Every time we
-    // modify something, we should be deleting an instruction, so we have not
-    // found any further information.
-    if (!visitedSinceLastMutation.insert(next).second) {
-      continue;
-    }
-
-    // First check if this is an instruction that is trivially dead. This can
-    // occur if we eliminate rr traffic resulting in dead projections and the
-    // like.
-    //
-    // If we delete, we first add all of our deleted instructions operands to
-    // the worklist and then remove all results (since we are going to delete
-    // the instruction).
-    if (auto *defInst = next->getDefiningInstruction()) {
-      if (isInstructionTriviallyDead(defInst)) {
-        assert(!ctx.assumingAtFixedPoint &&
-               "Assumed was at fixed point and recomputing state?!");
-        deleteAllDebugUses(defInst);
-        eraseInstruction(defInst);
-        madeChange = true;
-        ctx.verify();
-        continue;
-      }
-    }
-
-    // Otherwise, if we have a single value instruction (to be expanded later
-    // perhaps), try to visit that value recursively.
-    if (auto *svi = dyn_cast<SingleValueInstruction>(next)) {
-      bool madeSingleChange = visit(svi);
-      assert((!madeSingleChange || !ctx.assumingAtFixedPoint) &&
-             "Assumed was at fixed point and modified state?!");
-      madeChange |= madeSingleChange;
-      if (madeSingleChange) {
-        ctx.verify();
-      }
-      continue;
-    }
-  }
-
-  return madeChange;
-}
 
 //===----------------------------------------------------------------------===//
 //                            Top Level Entrypoint
@@ -179,7 +62,7 @@ struct SemanticARCOpts : SILFunctionTransform {
     }
 
     // Then process the worklist, performing peepholes.
-    bool madeChange = visitor.optimize();
+    bool eliminatedARCInst = visitor.optimize();
 
     // Now that we have seeded the map of phis to incoming values that could be
     // converted to guaranteed, ignoring the phi, try convert those phis to be
@@ -191,7 +74,7 @@ struct SemanticARCOpts : SILFunctionTransform {
     }
 
     // Otherwise, we only deleted instructions and did not touch phis.
-    if (madeChange)
+    if (eliminatedARCInst)
       invalidateAnalysis(SILAnalysis::InvalidationKind::Instructions);
   }
 };
